requirements.txt
# Core dependencies for the Gyroscopic Stabilized Quantum Engine
torch>=1.9.0
numpy>=1.21.0
scipy>=1.7.0
PyYAML>=5.4.0

# Optional dependencies for enhanced recording
pandas>=1.3.0
pyarrow>=5.0.0

# Development and testing
pytest>=6.0.0
pytest-cov>=2.12.0

# Network analysis (for spectral dimension estimation)
networkx>=2.6.0

# Optimization
cvxpy>=1.1.0